# Optimize reading dictionary pages